Removing session userID from code

+++ games/process/lfg.php
@@ -1,11 +1,8 @@
 <?
-	checkLogin(1);
-	
-	$userID = intval($_SESSION['userID']);
 	if (isset($_POST['update'])) {
-		$mysql->query("DELETE FROM lfg WHERE userID = $userID");
+		$mysql->query("DELETE FROM lfg WHERE userID = {$currentUser->userID}");
 		$lfgInsert = $mysql->prepare('INSERT INTO lfg SET userID = :userID, systemID = :systemID');
-		$lfgInsert->bindValue(':userID', $userID);
+		$lfgInsert->bindValue(':userID', $currentUser->userID);
 		$lfgInsert->bindParam(':systemID', $systemID);
 		foreach ($_POST['lfg'] as $systemID => $junk) $lfgInsert->execute();
